+++ service/src/main/kotlin/io/provenance/explorer/web/v2/GeneralController.kt
@@ -1,7 +1,6 @@
 package io.provenance.explorer.web.v2
 
 import io.provenance.explorer.domain.models.explorer.DateTruncGranularity
-import io.provenance.explorer.domain.models.explorer.GasStatistics
 import io.provenance.explorer.domain.models.explorer.Spotlight
 import io.provenance.explorer.service.ExplorerService
 import io.provenance.explorer.web.BaseController
@@ -16,6 +15,7 @@ import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
+import javax.validation.constraints.Min
 
 @Validated
 @RestController
@@ -29,13 +29,21 @@ class GeneralController(private val explorerService: ExplorerService) : BaseCont
 
     @ApiOperation("Returns gas statistics")
     @GetMapping("/gas/statistics")
-    fun gasStatistics(@RequestParam(required = true) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) fromDate: DateTime,
-                      @RequestParam(required = true) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) toDate: DateTime,
-                      @RequestParam(required = false) granularity: DateTruncGranularity?
-    ) =
-        ResponseEntity.ok(explorerService.getGasStatistics(fromDate, toDate, granularity))
+    fun gasStatistics(
+        @RequestParam(required = true) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) fromDate: DateTime,
+        @RequestParam(required = true) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) toDate: DateTime,
+        @RequestParam(required = false) granularity: DateTruncGranularity?
+    ) = ResponseEntity.ok(explorerService.getGasStatistics(fromDate, toDate, granularity))
 
     @ApiOperation("Returns the ID of the chain associated with the explorer instance")
     @GetMapping("/chain/id")
     fun getChainId(): ResponseEntity<String> = ResponseEntity.ok(explorerService.getChainId())
+
+    @ApiOperation("Returns statistics on min gas fees for the chain")
+    @GetMapping("/gas/fees/statstics")
+    fun getGasFeeStatistics(
+        @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) fromDate: DateTime?,
+        @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) toDate: DateTime?,
+        @RequestParam(required = false, defaultValue = "14") @Min(1) dayCount: Int
+    ) = ResponseEntity.ok(explorerService.getGasFeeStatistics(fromDate, toDate, dayCount))
 }